requirements.txt
Flask==0.10.1
SQLAlchemy==1.0.3
SQLAlchemy-FullText-Search==0.2.3
redis==2.10.5
celery==3.1.19
#Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Session==0.2
Flask-Cache==0.13.1
#Flask-SuperAdmin==1.7.1
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.9.6
alembic==0.7.6
amqp==1.4.6
anyjson==0.3.3
ujson==1.34
simplejson==3.8.1
argparse==1.2.1
billiard==3.3.0.20
blinker==1.3
itsdangerous==0.24
kombu==3.0.26
mysql-connector-python-rf==2.1.3
python-gcm==0.1.5
pytz==2015.4
#requests==2.7.0
wsgiref==0.1.2
xlrd==0.9.3
openpyxl==2.3.0-b1
MySQL-python==1.2.5
gevent==1.1rc2
pyasn1==0.1.9
pyopenssl==0.15.1
ndg-httpsclient==0.4.0
requests==2.9.1
uwsgi-tools==1.0.2
uwsgidecorators==1.1.0
uwsgicachetop==1.0.0
uwsgiFouine==1.7.dev0
uwsgi_cache==1.0
#uwsgi_metrics==1.1.5
uwsgi-stub==0.1.2
uwsgi-sloth==2.1.2
python-redis-lock==2.3.0
